feat(bottom-tab): Added bottom tab navigation with floating button
This is by far the most painful component I have ever made.
Due to Android's behavior of not handling child view when the view is
out of the parent view, I had to do a workaround.

Workaround is inspired by:
https://github.com/alex-melnyk/react-native-multibar/

The workaround is tweaked for personal needs, updated with the latest RN
libaries and features